+++ R/prepare.R
@@ -286,7 +286,6 @@ prepare_ceteris_paribus <- function(x, variables = NULL) {
 }
 
 prepare_feature_importance <- function(x, max_features = 10, show_boxplot = TRUE,
-                                       x_title = NULL,
                                        margin = 0.15, digits = 3, rounding_function = round,
                                        ...) {
   ### Return the object for the FeatureImportance plot ###
@@ -357,7 +356,6 @@ prepare_feature_importance <- function(x, max_features = 10, show_boxplot = TRUE
   ret$x <- new_x
   ret$m <- m
   ret$x_min_max <- min_max
-  ret$x_title <- ifelse(is.null(x_title), "drop-out loss", x_title)
   ret$desc <- data.frame(type = "desc",
                          text = gsub("\n","</br>", desc))